feature: membership 2023 prep
