+++ components/Forms/utils/formFields/membershipApplication.json
@@ -143,8 +143,7 @@
           "options": [
             "Design",
             "Development",
-            "Externals",
-            "Internal Growth",
+            "Management",
             "Marketing",
             "Research"
           ]
