+++ package.json
@@ -42,6 +42,6 @@
     "jest": "^27.3.1",
     "postcss": "^8.3.11",
     "react-test-renderer": "^17.0.2",
-    "tailwindcss": "^2.2.19"
+    "tailwindcss": "^3.3.3"
   }
 }
