+++ tailwind.config.js
@@ -1,12 +1,11 @@
 const plugin = require("tailwindcss/plugin");
-const defaultTheme = require('tailwindcss/defaultTheme');
+const defaultTheme = require("tailwindcss/defaultTheme");
 module.exports = {
-  purge: [
+  content: [
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
     "./sections/**/*.{js,ts,jsx,tsx}",
   ],
-  darkMode: false, // or 'media' or 'class'
   theme: {
     container: {
       center: true,
@@ -20,7 +19,7 @@ module.exports = {
         gray: "#80858B",
         blue: "#89CFF0",
         fontHeadingColor: "#20252B",
-        fontBodyColor: "#3B3E42"
+        fontBodyColor: "#3B3E42",
       },
       backgroundColor: {
         green: "#41BC9C",
@@ -35,8 +34,8 @@ module.exports = {
       },
     },
     screens: {
-      '2xs': '320px',
-      'xs': '360px',
+      "2xs": "320px",
+      xs: "360px",
       ...defaultTheme.screens,
     },
   },
@@ -44,7 +43,6 @@ module.exports = {
     extend: {},
   },
   plugins: [
-    require('@tailwindcss/line-clamp'),
     plugin(({ addBase, theme }) => {
       addBase({
         h1: { fontSize: theme("fontSize.h1") },
